ml/requirements.txt
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
flask==2.3.2
requests==2.31.0
psycopg2-binary==2.9.6
